pyproject.toml
[project]
name = "financial-agents-cicd"
version = "0.1.0"
description = "Financial agents with CI/CD pipeline for portfolio and market analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.13",
    "langchain-openai>=0.3.0",
    "langsmith>=0.2.0",
    "langgraph>=0.2.58",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "openevals>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "black>=24.0.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "evaluator: marks tests as offline evaluators (deselect with '-m \"not evaluator\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

